feat(api): Individual Container CRUD operations
This feature adds two basic container CRUD operations:

Inspecting a container with a GET request to /api/containers/id
Removing a container with a DELETE request to /api/containers/id
Under the hood I also changed the behind the scenes API requests so it now delivers consistent results by calling the same methods.

Oh! And there is more documentation in the Readme regarding the current state of the API.